pyproject.toml
[project]
name = "test-kube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "google-generativeai>=0.8.6",
    "greenlet>=3.3.1",
    "httpx>=0.24.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=0.3.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pandas>=3.0.0",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0",
    "pydantic-settings>=2.12.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=20.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "textblob>=0.19.0",
    "uvicorn>=0.40.0",
    "yfinance>=1.1.0",
]
